% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization.R
\name{mod_visualization_server}
\alias{mod_visualization_server}
\title{Visualization Server Module}
\usage{
mod_visualization_server(id, data)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{data}{A reactive data frame to use for plotting.}
}
\description{
Visualization Server Module
}
