% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization.R
\name{mod_visualization_ui}
\alias{mod_visualization_ui}
\title{Visualization UI Module}
\usage{
mod_visualization_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}
}
\value{
A UI definition for the visualization module.
}
\description{
Visualization UI Module
}
